Add login.html page alongside minor changes
In WebConfig, even though a lot of the mapping to /login, /login?logout=true, etc. is done under the hood, I've written the code out
in an explicit manner, so that it's easier to understand for others reading the code. It also shows that I understand what
the code does.

I've also renamed email-sheet.css to stylesheet.css as a lot of html templates use it now. Finally, I've added a .bright-color class,
that just gives the text an orange color (#fc941c).