Fix security issues with static files
The static files weren't showing because they were being served from
/static/, but I was trying to use them with the path /resources/** or /static/**
etc., but there are no /resources/ or /static/ subdirectories inside /static/.

Therefore, an initial solution was to use /**, however, this file path is recognised twice;
as the source root for static files, and for the root of the whole project.

Therefore, the above solution was not allowable. By creating subdirectories inside static,
we solved the problem by allowing for /css/**, /js/**, /images/**.